Input url and non-strict
Added the option to neglect the protocol/scheme in the input url. If
the script detects a valid V3 onion address, it will prepend 'http://',
while if it is anything else it will assume 'https://'. It also asks the
user to confirm if they use anything else than https for a non onion
address.

Secondly, I've changed the verbose behaviour when the strict variable
is turned off such that the output will clarify which part of the OMG
that are partially broken.